moved all vehicle stylesheets into new styles directory
