+++ src/components/vehicle/card.js
@@ -1,5 +1,6 @@
 import React, { useState } from "react";
-import "./card.css"
+// import "./card.css"
+import "../../styles/vehicles/card.css"
 
 const VehicleCard = props => {
