+++ src/components/vehicle/list.js
@@ -1,7 +1,8 @@
 import React, { useState, useEffect } from "react";
 import VehicleCard from "./vehicleCard"
 import VehicleManager from "../../api/dataManager";
-import "./list.css"
+// import "./list.css"
+import "../../styles/vehicles/list.css"
 import ModalWrapper from "../modal/modalWrapper"
 
 const VehiclesList = props => {
