+++ src/components/vehicle/modalAddVTForm.js
@@ -2,8 +2,10 @@
 
 import React, { useEffect, useState, useRef } from "react";
 import VehicleManager from "../../api/dataManager"
-import "./list.css";
-import "./modalAddForm.css";
+// import "./list.css";
+import "../../styles/vehicles/list.css"
+// import "./modalAddForm.css";
+import "../../styles/vehicles/addForm.css"
 
 const AddVehicleTypeForm = (props) => {
