+++ src/components/vehicle/vehicleCard.js
@@ -1,5 +1,6 @@
 import React, { useEffect } from "react";
-import "./searchCard.css"
+// import "./searchCard.css"
+import "../../styles/vehicles/searchCard.css"
 import NumberFormat from 'react-number-format';
 
 const VehicleCard = props => {
