+++ src/components/vehicle/vehicleDetailsModal.js
@@ -1,5 +1,6 @@
 import React from "react";
-import "./card.css"
+// import "./card.css"
+import "../../styles/vehicles/card.css"
 import "../employee/list.css"
 import "../employee/card.css"
