+++ src/styles/vehicles/addForm.css
@@ -1,0 +1,9 @@
+.modal-add--body {
+    overflow: scroll;
+    max-height: 79%;
+}
+
+.modal--input.msrp-div {
+    background-color: #FFF;
+    padding: 3px 0 3px 6px;
+}
