+++ src/styles/vehicles/card.css
@@ -1,0 +1,26 @@
+.vehicle-card--container {
+    border-bottom: 1px solid lightgray;
+    margin-bottom: 15px;
+}
+.vehicle-card--container:hover {
+    border-bottom: 2px solid rgba(54, 162, 235, 0.8);
+    cursor: pointer;
+}
+
+.vehicle-card--name {
+    font-family: 'Roboto', sans-serif;
+    font-weight: 400;
+    font-size: 18px;
+}
+
+.vehicle-card--numberSold {
+    font-family: 'Roboto', sans-serif;
+    font-weight: 300;
+    font-size: 16px;
+    margin-bottom: 5px;
+    color: rgb(102, 102, 102);
+}
+
+.label {
+    margin: 0;
+}
