+++ src/styles/vehicles/dashList.css
@@ -1,0 +1,16 @@
+.vehiclesContainer {
+    padding: 10px;
+    min-height: 250px;
+    min-width: 270px;
+    max-width: 270px;
+    max-height: 250px;
+    overflow: scroll;
+    -ms-overflow-style: none;  /* IE and Edge */
+    scrollbar-width: none;  /* Firefox */
+    color:'#33475B';
+}
+
+/* Hide scrollbar for Chrome/Safari/Opera */
+.vehiclesContainer::-webkit-scrollbar {
+    display: none;
+}
