+++ src/styles/vehicles/list.css
@@ -1,0 +1,134 @@
+.vehicles--container {
+    margin-left: 13%;
+}
+
+.vehicles--subContainer {
+    background-color: #F5F8FA;
+    box-shadow: 1px 3px 4px 1px #cacaca;
+    border-radius: 7px;
+    width: 95%;
+    height: 800px;
+    margin-top: 3%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.vehicles--header {
+    margin-top: 40px;
+    margin-bottom: 30px;
+    font-family: 'Roboto', sans-serif;
+    font-weight: 300;
+    font-size: 3rem;
+    color: #33475B;
+}
+
+.vehicles-searchBar {
+    width: 400px;
+    line-height: 35px;
+    border: 2px solid lightgray;
+    border-radius: 4px;
+    padding-left: 7px;
+}
+.vehicles-searchBar:focus {
+    outline: none;
+    border-color: #6a78d1;
+    box-shadow: 0 0 8px #6a78d1;
+}
+
+.searchResults.vehicles {
+    border: 2px solid lightgray;
+    box-shadow: 1px 5px 10px 1px lightgrey;
+    border-radius: 5px;
+    background-color: #fff;
+    height: 400px;
+    overflow: scroll;
+    -ms-overflow-style: none;
+    /* IE and Edge */
+    scrollbar-width: none;
+    /* Firefox */
+    width: 50%;
+    margin-top: 5%;
+}
+
+.addVehicleType--btn {
+    line-height: 40px;
+    width: 200px;
+    margin-top: 3.5%;
+    border-radius: 5px;
+    border: none;
+    background-color: #FF7A59;
+    color: #252222;
+    transition: transform .5s ease;
+}
+.addVehicleType--btn:hover {
+    transform: scale(1.07);
+}
+.addVehicleType--btn:focus {
+    outline: none;
+}
+
+.addVehicle--btn {
+    line-height: 40px;
+    width: 200px;
+    margin-top: 3.5%;
+    border-radius: 5px;
+    border: none;
+    background-color: #FF7A59;
+    color: #252222;
+    transition: transform .5s ease;
+}
+.addVehicle--btn:hover {
+    transform: scale(1.07);
+}
+.addVehicle--btn:focus {
+    outline: none;
+}
+
+.modal--form {
+    margin-top: 70px !important;
+    border-radius: 8px;
+    max-height: 750px;
+    overflow: scroll;
+}
+
+
+.modalBody {
+    border-radius: 8px;
+    display: flex !important;
+    flex-direction: column !important;
+    align-items: center !important;
+}
+
+.fieldset {
+    display: flex;
+    flex-direction: column;
+}
+
+.vehicleType--dropdown {
+    max-width: 200px;
+    max-height: 200px;
+    overflow: scroll;
+    border: 1px solid gray;
+    border-radius: 8px;
+    margin-top: 15px;
+    padding: 10px;
+}
+
+.vehicleType--select:hover {
+    cursor: pointer;
+    color: lightgray;
+}
+
+.modal--input {
+    line-height: 25px;
+    border: 2px solid gray;
+    border-radius: 4px;
+    padding-left: 7px;
+}
+
+.modal--input:focus {
+    outline: none;
+    /* border-color: #9ecaed;
+    box-shadow: 0 0 8px #9ecaed; */
+}
